Ignore spurious diffs in cpu resource requests/limits
There are multiple possible formats for the same semantic value (eg. 50m vs 0.05)
for cpu resources, and the k8s server may return one even if you apply the other.
This creates a diff that can't be resolved by applying.

We solve this problem by ignoring the diff if the values have the same meaning.
To facilitate this we introduce a generic method for adding 'tolerations' to the diffs.

Note that we need to do this inside diff() since diff_list() uses diff to check for equality
and won't behave correctly if we tried to strip the tolerable diffs later.

I've used globbing of paths for simplicity, as it allows you to easily say "anything ending with"
or "all paths under", etc. without resorting to full regex, which I think would be less readable
in most cases.

Note there are also a few minor edits to make this behave correctly - eg. normalize needs to normalize
floats as well as ints, and diff_lists needs to give '' as an empty path rather than None so tolerations
can be matched correctly.

Fixes #33